refactor: test/support 하위 폴더 추가
