+++ backend/src/test/java/backend/mulkkam/auth/domain/OauthAccountTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import backend.mulkkam.common.exception.CommonException;
 import backend.mulkkam.member.domain.Member;
-import backend.mulkkam.support.MemberFixtureBuilder;
+import backend.mulkkam.support.fixture.MemberFixtureBuilder;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
