+++ backend/src/test/java/backend/mulkkam/intake/repository/TargetTargetAmountSnapshotRepositoryTest.java
@@ -6,7 +6,7 @@ import backend.mulkkam.intake.domain.TargetAmountSnapshot;
 import backend.mulkkam.member.domain.Member;
 import backend.mulkkam.member.domain.vo.TargetAmount;
 import backend.mulkkam.member.repository.MemberRepository;
-import backend.mulkkam.support.MemberFixtureBuilder;
+import backend.mulkkam.support.fixture.MemberFixtureBuilder;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
