+++ backend/src/test/java/backend/mulkkam/notification/service/SuggestionNotificationIntegrationServiceTest.java
@@ -14,10 +14,10 @@ import backend.mulkkam.member.repository.MemberRepository;
 import backend.mulkkam.notification.domain.Notification;
 import backend.mulkkam.notification.domain.SuggestionNotification;
 import backend.mulkkam.notification.repository.SuggestionNotificationRepository;
-import backend.mulkkam.support.MemberFixtureBuilder;
-import backend.mulkkam.support.NotificationFixtureBuilder;
-import backend.mulkkam.support.ServiceIntegrationTest;
-import backend.mulkkam.support.SuggestionNotificationFixtureBuilder;
+import backend.mulkkam.support.fixture.MemberFixtureBuilder;
+import backend.mulkkam.support.fixture.NotificationFixtureBuilder;
+import backend.mulkkam.support.service.ServiceIntegrationTest;
+import backend.mulkkam.support.fixture.SuggestionNotificationFixtureBuilder;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
