+++ backend/src/test/java/backend/mulkkam/support/fixture/AccountRefreshTokenFixtureBuilder.java
@@ -1,4 +1,4 @@
-package backend.mulkkam.support;
+package backend.mulkkam.support.fixture;
 
 import backend.mulkkam.auth.domain.AccountRefreshToken;
 import backend.mulkkam.auth.domain.OauthAccount;
