+++ backend/src/test/java/backend/mulkkam/support/fixture/AppMinimumVersionFixtureBuilder.java
@@ -1,4 +1,4 @@
-package backend.mulkkam.support;
+package backend.mulkkam.support.fixture;
 
 import backend.mulkkam.version.domain.AppMinimumVersion;
 import java.time.LocalDateTime;
