+++ backend/src/test/java/backend/mulkkam/support/fixture/CupFixtureBuilder.java
@@ -1,4 +1,4 @@
-package backend.mulkkam.support;
+package backend.mulkkam.support.fixture;
 
 import backend.mulkkam.cup.domain.Cup;
 import backend.mulkkam.cup.domain.CupEmoji;
