+++ backend/src/test/java/backend/mulkkam/support/fixture/NotificationFixtureBuilder.java
@@ -1,4 +1,4 @@
-package backend.mulkkam.support;
+package backend.mulkkam.support.fixture;
 
 import backend.mulkkam.member.domain.Member;
 import backend.mulkkam.notification.domain.Notification;
