+++ backend/src/test/java/backend/mulkkam/support/fixture/SuggestionNotificationFixtureBuilder.java
@@ -1,4 +1,4 @@
-package backend.mulkkam.support;
+package backend.mulkkam.support.fixture;
 
 import backend.mulkkam.notification.domain.Notification;
 import backend.mulkkam.notification.domain.SuggestionNotification;
