+++ backend/src/test/java/backend/mulkkam/support/fixture/TargetAmountSnapshotFixtureBuilder.java
@@ -1,4 +1,4 @@
-package backend.mulkkam.support;
+package backend.mulkkam.support.fixture;
 
 import backend.mulkkam.intake.domain.TargetAmountSnapshot;
 import backend.mulkkam.member.domain.Member;
