+++ backend/src/test/java/backend/mulkkam/support/service/ServiceIntegrationTest.java
@@ -1,5 +1,6 @@
-package backend.mulkkam.support;
+package backend.mulkkam.support.service;
 
+import backend.mulkkam.support.DatabaseCleaner;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
